Arrow draw error while deleting table fixed.
