+++ silzila-frontend/src/App.css
@@ -1,8 +1,8 @@
 .App {
-    text-align: center;
+	text-align: center;
 }
 
 body {
-    margin: 0;
-    padding: 0;
+	margin: 0;
+	padding: 0;
 }
