+++ silzila-frontend/src/Components/DataConnection/DashBoardList.js
@@ -1,23 +1,23 @@
 import React from "react";
 
 const DashBoardList = () => {
-    return (
-        <div className="dashboardsContainer">
-            <div className="containersHead">
-                <div className="containerTitle">DashBoard</div>
+	return (
+		<div className="dashboardsContainer">
+			<div className="containersHead">
+				<div className="containerTitle">DashBoard</div>
 
-                <input
-                    className="containerButton"
-                    type="button"
-                    value="New"
-                    // onClick={(e) => {
-                    //     handleOpen();
-                    //     handleMode(e);
-                    // }}
-                />
-            </div>
-        </div>
-    );
+				<input
+					className="containerButton"
+					type="button"
+					value="New"
+					// onClick={(e) => {
+					//     handleOpen();
+					//     handleMode(e);
+					// }}
+				/>
+			</div>
+		</div>
+	);
 };
 
 export default DashBoardList;
