+++ silzila-frontend/src/Components/DataSet/BottomBar.js
@@ -1,7 +1,8 @@
 import React from "react";
 
 const BottomBar = () => {
-    return <div className="bottomBar">BottomBar</div>;
+	// TODO Parse dataSet info and make api call to create Dataset
+	return <div className="bottomBar">BottomBar</div>;
 };
 
 export default BottomBar;
