+++ silzila-frontend/src/Components/DataSet/Tables.js
@@ -7,123 +7,124 @@ import CanvasTableColumns from "./CanvasTableColumns";
 import { useXarrow } from "react-xarrows";
 
 const Tables = ({
-    // props
-    tableData,
+	// props
+	tableData,
 
-    // state
-    tableList,
-    tempTable,
-    arrows,
-    actions,
+	// state
+	tableList,
+	tempTable,
+	arrows,
+	actions,
 
-    // dipatch
+	// dipatch
 }) => {
-    const dragRef = useRef();
-    const updateXarrow = useXarrow();
+	const dragRef = useRef();
+	const updateXarrow = useXarrow();
 
-    let tableItems = tableData.columns;
-    let tableId = tableData.tableName;
-    let tableTitle = tableData.tableName;
-    let alias = tableData.alias;
+	let tableItems = tableData.columns;
+	let tableId = tableData.tableName;
+	let tableTitle = tableData.tableName;
+	let alias = tableData.alias;
 
-    const [inputField, setInputField] = useState(false);
-    const [newName, setNewName] = useState("");
-    const [selectedTable, setSelectedTable] = useState();
-    const [open, setOpen] = useState(false);
+	const [inputField, setInputField] = useState(false);
+	const [newName, setNewName] = useState("");
+	const [selectedTable, setSelectedTable] = useState();
+	const [open, setOpen] = useState(false);
 
-    const changeTableName = (tableName) => {
-        const newTable = [...tempTable].map((tab) => {
-            if (tab.alias === tableName) {
-                tab.alias = newName;
-            }
-            return tab;
-        });
-        setNewName("");
-        setInputField(false);
-    };
+	const changeTableName = (tableName) => {
+		const newTable = [...tempTable].map((tab) => {
+			if (tab.alias === tableName) {
+				tab.alias = newName;
+			}
+			return tab;
+		});
+		setNewName("");
+		setInputField(false);
+	};
 
-    return (
-        <div>
-            <Draggable
-                ref={dragRef}
-                handle={`#${tableId}`}
-                bounds="#canvasTableArea"
-                onDrag={updateXarrow}
-                onStop={updateXarrow}
-            >
-                <div className="draggableBox">
-                    {inputField ? (
-                        <div>
-                            <TextField
-                                variant="standard"
-                                id="name"
-                                className="inputArea"
-                                value={newName}
-                                onChange={(e) => {
-                                    e.preventDefault();
-                                    setNewName(e.target.value);
-                                }}
-                            />
-                            <Button
-                                type="button"
-                                className="button_Ok"
-                                onClick={() => changeTableName(alias)}
-                            >
-                                Ok
-                            </Button>
-                        </div>
-                    ) : (
-                        <div className="draggableBoxTitle" id={tableId}>
-                            <div
-                                onDoubleClick={() => {
-                                    setInputField(true);
-                                    setNewName(alias);
-                                    setTimeout(() => {
-                                        // selectText();
-                                    }, 10);
-                                }}
-                            >
-                                {alias}
-                            </div>
-                            <div>
-                                <MoreVertIcon
-                                    style={{ float: "right" }}
-                                    onClick={() => {
-                                        setSelectedTable(tableTitle);
-                                        setOpen(true);
-                                    }}
-                                />
-                            </div>
-                        </div>
-                    )}
-                    {tableItems.map((item, index) => {
-                        return (
-                            // <div>{item.column_name}</div>
-                            <CanvasTableColumns
-                                columnName={item.column_name}
-                                tableName={tableTitle}
-                                itemType={item.data_type}
-                                index={index}
-                                handler="right"
-                                itemId={item.uid}
-                                tableId={tableId}
-                                dragRef={dragRef}
-                            />
-                        );
-                    })}
-                </div>
-            </Draggable>
-        </div>
-    );
+	return (
+		<div>
+			<Draggable
+				ref={dragRef}
+				handle={`#${tableId}`}
+				bounds="#canvasTableArea"
+				onDrag={updateXarrow}
+				onStop={updateXarrow}
+			>
+				<div className="draggableBox" ref={dragRef}>
+					{inputField ? (
+						<div>
+							<TextField
+								variant="standard"
+								id="name"
+								className="inputArea"
+								value={newName}
+								onChange={(e) => {
+									e.preventDefault();
+									setNewName(e.target.value);
+								}}
+							/>
+							<Button
+								type="button"
+								className="button_Ok"
+								onClick={() => changeTableName(alias)}
+							>
+								Ok
+							</Button>
+						</div>
+					) : (
+						<div className="draggableBoxTitle" id={tableId}>
+							<div
+								onDoubleClick={() => {
+									setInputField(true);
+									setNewName(alias);
+									setTimeout(() => {
+										// selectText();
+									}, 10);
+								}}
+							>
+								{alias}
+							</div>
+							<div>
+								<MoreVertIcon
+									style={{ float: "right" }}
+									onClick={() => {
+										setSelectedTable(tableTitle);
+										setOpen(true);
+									}}
+								/>
+							</div>
+						</div>
+					)}
+					{tableItems.map((item, index) => {
+						return (
+							// <div>{item.column_name}</div>
+							<CanvasTableColumns
+								key={item.uid}
+								columnName={item.column_name}
+								tableName={tableTitle}
+								itemType={item.data_type}
+								index={index}
+								handler="right"
+								itemId={item.uid}
+								tableId={tableId}
+								dragRef={dragRef}
+							/>
+						);
+					})}
+				</div>
+			</Draggable>
+		</div>
+	);
 };
 
 const mapStateToProps = (state) => {
-    return {
-        tableList: state.dataSetState.tables,
-        tempTable: state.dataSetState.tempTable,
-        arrows: state.dataSetState.arrows,
-        actions: state.dataSetState.actions,
-    };
+	return {
+		tableList: state.dataSetState.tables,
+		tempTable: state.dataSetState.tempTable,
+		arrows: state.dataSetState.arrows,
+		actions: state.dataSetState.actions,
+	};
 };
 
 export default connect(mapStateToProps, null)(Tables);
