+++ silzila-frontend/src/redux/Dataset/datasetActions.js
@@ -26,7 +26,7 @@ export const removeArrows = (pl) => {
 	return { type: "REMOVE_ARROWS", payload: pl };
 };
 export const resetState = () => {
-    return { type: "RESET_STATE" };
+	return { type: "RESET_STATE" };
 };
 
 // =============================================================================
