+++ silzila-frontend/src/redux/UserInfo/isLoggedActions.js
@@ -1,4 +1,4 @@
 export const userAuthentication = (payload) => {
-    console.log("User Authentication");
-    return { type: "USER_AUTHENTICATED", payload: payload };
+	console.log("User Authentication");
+	return { type: "USER_AUTHENTICATED", payload: payload };
 };
