+++ silzila-frontend/src/redux/index.js
@@ -3,8 +3,8 @@ import DataSetReducer from "./Dataset/dataset";
 import loggedReducer from "./UserInfo/isLogged";
 
 const allReducers = combineReducers({
-    isLogged: loggedReducer,
-    dataSetState: DataSetReducer,
+	isLogged: loggedReducer,
+	dataSetState: DataSetReducer,
 });
 
 export default allReducers;
